Merge $XML_CATALOG_FILES with settings.
On most Linux boxes the $XML_CATALOG_FILES already contains valid paths.

User settings are not mandatory then, but Exalt replaces the valid environment
variables with its own settings which may not be correct. This may prevent
lxml to find catalog files and makes it crucial to set up the paths in Exalt's
settings.

As lxml is used by other ST packages such as XPath, too, this may not be desirable.

Therefore catalog files from Exalt.sublime-settings should rather be merged with
existing entries in $XML_CATALOG_FILES instead of replacing them.

This is what this commit / PR does.